requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# CLI interface
rich>=13.7.0
prompt-toolkit>=3.0.43
click>=8.1.7

# Git operations
GitPython>=3.1.40

# LLM integration
litellm>=1.0.0

# Data handling
pydantic>=2.5.0
dataclasses-json>=0.6.3

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
types-requests>=2.31.0

# Additional utilities
PyYAML>=6.0.1
urllib3>=2.1.0